% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEllipse.R
\name{addEllipse}
\alias{addEllipse}
\title{Ellipse}
\usage{
addEllipse(means, varCov)
}
\arguments{
\item{means}{a numeric vector of length 2. The center of the ellipse.}

\item{varCov}{a numeric value (the value will be transformed in 2x2 matrix)
or 2x2 matrix. The direction of the ellipse.}
}
\value{
plot of the ellipses and their centers.
}
\description{
Draw an ellipse on top of a plot.
}
