% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupGaussianData.R
\name{groupGaussianData}
\alias{groupGaussianData}
\title{EM for gaussian mixtures.}
\usage{
groupGaussianData(
  D,
  K,
  modelType = "diagonal",
  randomInit = 10,
  endIter = FALSE,
  nbIter = 100
)
}
\arguments{
\item{D}{A matrix, or data.frame containing continuous variable(s)}

\item{K}{The desired number of clusters}

\item{modelType}{a character, defining a family of gaussian models. 3 models
with free proportions are implemented : "general", "diagonal" and "spherical".}

\item{randomInit}{a numeric value. The number of random initializations of the EM algorithm.
The default is randomInit = 10.}

\item{endIter}{a boolean, specifying the stopping rule for the EM algorithm.
If endIter = TRUE, the EM alogrithm is stopped using the defined number of
iterations. If FALSE (default), the EM is stopped, if the incomplete likelihood converge.}

\item{nbIter}{a numeric value. The number of iterations desired to be performed by the EM algorithm.
This parameter is used only when endIter = TRUE.}
}
\value{
"gaussianClust" object, which contains the clustering results
}
\description{
This function performs a clustering of a univariate or multivariate continuous futures by using a gaussian mixture model.
}
\examples{
data("iris")
mod_gaussian <- groupGaussianData(iris[-5], 3, modelType = "diagonal", endIter = FALSE)
## not run
#summaryResults(mod_gaussian)
#plotResults(mod_gaussian, iris)

}
