% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMixData.R
\name{groupMixData}
\alias{groupMixData}
\title{EM for mixed futures, Gaussian-Mutinomial mixture models.}
\usage{
groupMixData(
  D,
  K,
  modelType = "diagonal",
  randomInit = 10,
  endIter = FALSE,
  nbIter = 100
)
}
\arguments{
\item{D}{A Dataframe. The dataset to perform the clustering on.}

\item{K}{The desired number of clusters}

\item{modelType}{a character, defining a family of gaussian models. 3 models
with free proportions are implemented : "general", "diagonal" and "spherical".}

\item{randomInit}{a numeric value. The number of random initializations of the EM algorithm.
The default is randomInit = 10.}

\item{endIter}{a boolean, specifying the stopping rule for the EM algorithm.
If endIter = TRUE, the EM alogrithm is stopped using the defined number of
iterations. If FALSE (default), the EM is stopped, if the incomplete likelihood converge.}

\item{nbIter}{a numeric value. The number of iterations desired to be performed by the EM algorithm.
This parameter is used only when endIter = TRUE.}
}
\description{
Clustering of a dataset with mix variables (gaussian and multinomial) and unknown labels. The quantitative features are assumed to be mixtures of gaussian distributions. The categorical features are assumed to be
mixtures of multinomial distributions. The quantitative and qualitative variables are assumed to be
independent with respect to the clusters.
}
\examples{
data("ToothGrowth")
# mix data, 1 categorical and 2 continuous variables
str(ToothGrowth)
mod_mix <-groupMixData(ToothGrowth, 2, modelType = "spherical")
## not run
#summaryResults(mod_mix)
#plotResults(mod_mix, ToothGrowth)
}
