% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.R
\name{criterionBIC}
\alias{criterionBIC}
\alias{criterionAIC}
\title{Model selection criteria}
\usage{
criterionBIC(loglik, nb_params, nb_obs)

criterionAIC(loglik, nb_params)
}
\arguments{
\item{loglik}{a numeric value. The log-liklihood of the model.}

\item{nb_params}{a numeric value. The number of parameters of in the model.}

\item{nb_obs}{a numeric value. The number of observations in the dataset.}
}
\description{
Functions for computing the model selection criteria.
}
\section{Functions}{
\itemize{
\item \code{criterionBIC}: Compute the BIC criterion

\item \code{criterionAIC}: Compute the AIC criterion
}}

