% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryResults.R
\name{summaryResults}
\alias{summaryResults}
\title{Summary of the clustering results}
\usage{
summaryResults(object)
}
\arguments{
\item{object}{mixclust object. 3 possible objects : "mixClust",
gaussianClust", "multinomialClust".}
}
\value{
nice print.
}
\description{
Pretty print of the clustering results.
}
\examples{
data("iris")
mod_gaussian <- groupGaussianData(iris[-5], 3, modelType = "diagonal", endIter = FALSE)
summaryResults(mod_gaussian)

}
