% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMultinomialData.R
\name{groupMultinomialData}
\alias{groupMultinomialData}
\title{EM for Multinomial mixtures}
\usage{
groupMultinomialData(D, K, randomInit = 10, endIter = FALSE, nbIter = 100)
}
\arguments{
\item{D}{A data.frame, or table, containing only categorical variable(s)}

\item{K}{The desired number of clusters}

\item{randomInit}{a numeric value. The number of random initializations of the EM algorithm.
The default is randomInit = 10.}

\item{endIter}{a boolean, specifying the stopping rule for the EM algorithm.
If endIter = TRUE, the EM alogrithm is stopped using the defined number of
iterations. If FALSE (default), the EM is stopped, if the incomplete likelihood converge.}

\item{nbIter}{a numeric value. The number of iterations desired to be performed by the EM algorithm.
This parameter is used only when endIter = TRUE.}
}
\value{
"multinomialClust" object, which contains the clustering results
}
\description{
This function performs clustering on a dataset composed by (d >= 1) independent categorical
variable(s) given the latent variable.
}
\examples{
data("HairEyeColor")
mod_mult <- groupMultinomialData(HairEyeColor, 3, endIter = TRUE)
## not run
#summaryResults(mod_mult)
#plotResults(mod_mult, "HairEyeColor")
}
