% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Visualization of the obtained clusters}
\usage{
plotResults(object, D)
}
\arguments{
\item{object}{mixclust object. 3 possible objects : "mixClust",
gaussianClust", "multinomialClust".}

\item{D}{The dataset containing quantitative variable and/or qualitative variable.}
}
\description{
Plot the results of "mixClust", "gaussianClust" and "multinomialClust" objects.
}
